+++ R/archeofrag.gui.R
@@ -1,10 +1,14 @@
 
-archeofrag.gui <- function(){
+archeofrag.gui <- function(n.cores=NULL){
   shiny::addResourcePath('www', system.file('www', package = 'archeofrag.gui'))
-  doParallel::registerDoParallel() #  half of available cores parallel::detectCores()
+  
+  if( ! is.null(n.cores)){
+    if(n.cores <= parallel::detectCores()){
+      doParallel::registerDoParallel(cl = n.cores) 
+    } else{warning(paste(parallel::detectCores(), "cores available, can't use", n.cores, "cores.")) }
+  } else{
+    doParallel::registerDoParallel() # register half of available cores
+  }
   # foreach::getDoParWorkers() 
   shinyApp(ui = ui, server = server)
 }
-
-
-