+++ app.R
@@ -1,7 +1,3 @@
 pkgload::load_all(export_all = FALSE, helpers = FALSE, attach_testthat = FALSE)
 
 archeofrag.gui::archeofrag.gui()
-
-
-
-